Labfiles/09-build-remote-agents-with-a2a/python/requirements.txt
python-dotenv
httpx
azure-identity
uvicorn
starlette
sse-starlette
fastapi
azure-ai-projects 
azure-ai-agents 
a2a-sdk==0.3.26
